version2/fastapi_app/pyproject.toml
[tool.poetry]
name = "src"
version = "2023.04.08"
description = "restapi for climate data"
authors = ["darrida <darrida.py@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
psycopg = {extras = ["binary"], version = "^3.1.4"}
psycopg-pool = "^3.1.4"
loguru = "^0.6.0"
nicegui = "^1.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
sqlalchemy = "^2.0.7"
email-validator = "^1.3.1"
bcrypt = "^4.0.1"
requests = "^2.28.2"
itsdangerous = "^2.1.2"

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
extend-exclude = [".venv"]
extend-select = ["S", "I", "C4"]  # bandit (S), isort (I), comprehensions (C4), type-checking (TCH)

[tool.ruff.isort]
extra-standard-library = []
known-first-party = ["src"]
known-third-party = []
known-local-folder = []
# relative-imports-order = "closest-to-furthest"
# required-imports = ["from __future__ import annotations"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
